feat: create abstract wallet scheduler
